% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{addClinicalCovariates}
\alias{addClinicalCovariates}
\title{Add clinical covariates to chm}
\usage{
addClinicalCovariates(chm, study, cohort)
}
\arguments{
\item{chm}{NG-CHM}

\item{study}{The TCGA study identifier(s) of the data samples}

\item{cohort}{The participant identifiers included in the NGCHM}
}
\description{
Add clinical covariates to chm
}

