% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{defineClinicalCovariate}
\alias{defineClinicalCovariate}
\title{Define a clinical covariate for use in NGCHMs}
\usage{
defineClinicalCovariate(fullName, columnName, studies)
}
\arguments{
\item{fullName}{Display name of covariate}

\item{columnName}{Name of covariate in ISB CGC clinical table}

\item{studies}{Vector of TCGA studies for which covariate is valid. NULL (default) is valid for all.}
}
\description{
Define a clinical covariate for use in NGCHMs
}
\seealso{
removeClinicalCovariate
}

