% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{demoCHM}
\alias{demoCHM}
\title{Create a canned NG-CHM from ISB CGC cloud data}
\usage{
demoCHM(study = "prad", authority = "Vogelstein")
}
\arguments{
\item{study}{Include samples from the specified study (or studies).}

\item{authority}{Include genes deemed 'cancer interesting' by the specified authority (or authorities).}
}
\value{
The generated NG-CHM
}
\description{
This function creates a canned gene expression NG-CHM for the given study and gene authority using data
from the ISB CGC cloud pilot.  If the 'viewer' option is defined, it will be used to display the NG-CHM.
}
\seealso{
exprCHM
}

