% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{exprCHM}
\alias{exprCHM}
\title{Create a canned NG-CHM from ISB CGC cloud data}
\usage{
exprCHM(name, cohort, genes, caption = NULL, cbioStudy = NULL)
}
\arguments{
\item{name}{Name the generated NG-CHM 'name'.}

\item{cohort}{Include samples from the specified vector of participant ids.}

\item{genes}{Include genes from the specified vector of HGNC symbols.}

\item{caption}{An informative caption}

\item{cbioStudy}{cBio study identifier to use for cBio linkouts}
}
\value{
The generated NG-CHM
}
\description{
This function creates a canned gene expression NG-CHM for the given participant cohort and gene list using data
from the ISB CGC cloud pilot.  If the 'viewer' option is defined, it will be used to display the NG-CHM.
The generated NG-CHM will contain three layers: row-centered, Z-normalized, and original.
}
\seealso{
getExpressionData

demoCHM
}

