% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{getClinicalData}
\alias{getClinicalData}
\title{Get the specified columns from the clinical data for the specified cohort}
\usage{
getClinicalData(cohort, columns)
}
\arguments{
\item{cohort}{Include samples from the specified vector of participant ids.}

\item{columns}{Include columns from the specified vector of clinical table column names.}
}
\value{
A list of column vectors.
}
\description{
Get the specified columns from the clinical data for the specified cohort
}

