% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{getExpressionData}
\alias{getExpressionData}
\title{Get gene expression data from the ISB CGC cloud}
\usage{
getExpressionData(cohort, genes)
}
\arguments{
\item{cohort}{Include samples from the specified vector of participant ids.}

\item{genes}{Include genes from the specified vector of HGNC symbols.}
}
\value{
A numeric matrix of genes (rows) by samples (columns).
}
\description{
This function downloads gene expression data from the ISB CGC cloud as
a gene by sample matrix.  The data will be transformed using ASINH.
}

