% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{getMutationData}
\alias{getMutationData}
\title{Get mutation data.}
\usage{
getMutationData(cohort, genes)
}
\arguments{
\item{cohort}{Include samples from the specified vector of participant ids.}

\item{genes}{Include mutations in the specified vector of Hugo_Symbols.}
}
\value{
A list of column vectors.
}
\description{
Get mutation data.
}

