% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{removeClinicalCovariate}
\alias{removeClinicalCovariate}
\title{Remove a clinical covariate definition}
\usage{
removeClinicalCovariate(name)
}
\arguments{
\item{name}{Name of covariate to remove}
}
\description{
Remove a clinical covariate definition
}
\seealso{
defineClinicalCovariate
}

